requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
Babel==2.15.0
dj-database-url==2.1.0
Django==3.2.1
django-cors-headers==3.14.0
django-extra-views==0.14.0
django-haystack==3.3b2
django-oscar==3.2.4
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
factory-boy==3.2.1
Faker==25.3.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.17.3
packaging==23.2
phonenumbers==8.13.37
Pillow==9.4.0
psycopg2==2.9.7
psycopg2-binary==2.9.9
purl==1.6
pycodestyle==2.10.0
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.7
uritemplate==4.1.1
whitenoise==6.5.0
